Remove Filter Button from Evaluation Side Panel
Ticket: #772
