+++ src/app/evaluations/components/EvaluationMainContent.tsx
@@ -1,0 +1,48 @@
+import React from "react";
+import { MainComponentHeader } from "../../shared/components/MainComponentHeader";
+
+import Box from "@mui/material/Box";
+import { useTranslation } from "react-i18next";
+import { DivisionToken } from "../model/Evaluations.model";
+import { EvaluationDivisionContainer } from "./EvaluationDivisionContainer";
+import { useEvaluationContentComponent } from "./evaluationsContentUseCases";
+// import { useTranslation } from "react-i18next";
+interface EvaluationMainContentProps {
+    heading: string;
+}
+
+export function EvaluationMainContent({}: EvaluationMainContentProps): JSX.Element {
+    const { model, operations } = useEvaluationContentComponent();
+    const { t } = useTranslation(["ExplanationPage"]);
+    return (
+        <>
+            <MainComponentHeader heading={model.heading}></MainComponentHeader>
+            <Box
+                style={{
+                    overflowY: "scroll",
+                }}
+            >
+                {!model.loading &&
+                    Object.keys(model.evaluationsData)
+                        .filter((value) => operations.showDivision(value))
+                        .map((division) => (
+                            <EvaluationDivisionContainer
+                                key={division}
+                                title={t(division)} // Updated line
+                                divisionData={
+                                    model.evaluationsData[
+                                        division as DivisionToken
+                                    ]
+                                }
+                                downloadGraphButtonText={
+                                    model.downloadGraphButtonText
+                                }
+                                downloadDataButtonText={
+                                    model.downloadDataButtonText
+                                }
+                            />
+                        ))}
+            </Box>
+        </>
+    );
+}
