+++ src/app/evaluations/components/EvaluationSideContent.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { FilterContainerComponent } from "./FilterContainerComponent";
+import { useEvaluationSideComponent } from "./evaluationsSideUseCases";
+
+export function EvaluationSideContent(): JSX.Element {
+    const { model } = useEvaluationSideComponent();
+
+    return (
+        <FilterContainerComponent
+            selectionConfig={model.selectionConfig}
+            howToHeading={model.howToHeading}
+            howToContent={model.howto}
+        />
+    );
+}
