+++ src/app/evaluations/components/FilterContainerComponent.tsx
@@ -1,55 +1,25 @@
+import { Box, Tooltip } from "@mui/material";
+import Markdown from "markdown-to-jsx";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Box, Button, Tooltip } from "@mui/material";
-import FilterListIcon from "@mui/icons-material/FilterList";
-import Markdown from "markdown-to-jsx";
 import { ZNAccordion } from "../../shared/components/accordion/ZNAccordion";
-import {
-    FilterSelection,
-    SelectionFilterConfig,
-} from "../model/Evaluations.model";
+import { SelectionFilterConfig } from "../model/Evaluations.model";
 import { FilterMultiSelectionComponent } from "./FilterMultiSelectionComponent";
 
 type FilterContainerComponentProps = {
     selectionConfig: SelectionFilterConfig[];
-    searchButtonText: string;
     howToHeading: string;
     howToContent: string;
-    handleSearchBtnClick: (filter: FilterSelection) => void;
 };
 
 export function FilterContainerComponent({
     selectionConfig,
-    searchButtonText,
     howToContent,
     howToHeading,
-    handleSearchBtnClick,
 }: FilterContainerComponentProps): JSX.Element {
-    const { t } = useTranslation(["ExplanationPage"]);
-    const { i18n } = useTranslation();
+    const { t, i18n } = useTranslation(["ExplanationPage"]);
 
     const [tooltipOpen, setTooltipOpen] = useState<string | null>(null);
-
-    const initialFilters: FilterSelection = {
-        matrix: [],
-        productionType: [],
-        diagramType: [],
-        category: [],
-        microorganism: [],
-        division: [],
-    };
-
-    const gatherFilters = (): FilterSelection => {
-        return selectionConfig.reduce((filters, config) => {
-            filters[config.id as keyof FilterSelection] = config.selectedItems;
-            return filters;
-        }, initialFilters);
-    };
-
-    const handleSearch = (): void => {
-        const updatedFilters = gatherFilters();
-        handleSearchBtnClick(updatedFilters);
-    };
 
     return (
         <Box
@@ -96,24 +66,6 @@ export function FilterContainerComponent({
                     </Tooltip>
                 );
             })}
-            <Box
-                sx={{
-                    display: "flex",
-                    flexDirection: "row",
-                    padding: 1,
-                    justifyContent: "space-between",
-                    gap: 2,
-                }}
-            >
-                <Button
-                    variant="outlined"
-                    onClick={handleSearch}
-                    startIcon={<FilterListIcon />}
-                    sx={{ margin: "8px" }}
-                >
-                    {searchButtonText}
-                </Button>
-            </Box>
 
             <ZNAccordion
                 key="howTo"
