+++ src/app/evaluations/pages/evaluationsPageUseCases.ts
@@ -1,0 +1,19 @@
+import { useTranslation } from "react-i18next";
+
+const useEvaluationPageComponent = (): {
+    model: {
+        mainHeading: string;
+        sideBarTitle: string;
+    };
+} => {
+    const { t } = useTranslation(["ExplanationPage"]);
+
+    return {
+        model: {
+            mainHeading: t("Heading"),
+            sideBarTitle: t("Filter_Settings"),
+        },
+    };
+};
+
+export { useEvaluationPageComponent };
